requirements.txt
streamlit==1.28.0
pandas==2.0.3
gspread==5.10.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0